requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-classic
langchain-chroma
langchain-huggingface
langchain-google-genai
python-dotenv
beautifulsoup4
requests
pysqlite3-binary
